added remove uploaded files on controller failure

+++ backend/controllers/partners.controller.js
@@ -147,9 +147,11 @@ exports.createPartner = async (req, res) => {
     }
 
   } catch (err) {
-    req.files.forEach(el=>{
-      fs.unlinkSync(el.path)
-    })
+    if (typeof req.files != 'undefined') {
+      req.files.forEach(el => {
+        fs.unlinkSync(el.path)
+      })
+    }
     console.log(err)
     res.send(err)
   }
